fanciz Learning Project -Jpa Persistence Context Test

+++ src/main/java/com/jktoy/fancizToyProject/repository/TeamInfoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TeamInfoRepository extends JpaRepository<TeamInfo, Long> {
+public interface TeamInfoRepository extends JpaRepository<TeamInfo, Integer> {
 
 }
